+++ src/main/java/name/divinityunbound/block/entity/SpaceTimeFurnaceBlockEntity.java
@@ -1,0 +1,41 @@
+package name.divinityunbound.block.entity;
+
+import name.divinityunbound.screen.SpaceTimeFurnaceScreenHandler;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.Blocks;
+import net.minecraft.block.entity.AbstractFurnaceBlockEntity;
+import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.item.Items;
+import net.minecraft.recipe.RecipeType;
+import net.minecraft.screen.ScreenHandler;
+import net.minecraft.text.Text;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
+
+import java.util.Map;
+
+public class SpaceTimeFurnaceBlockEntity extends AbstractFurnaceBlockEntity {
+
+    public SpaceTimeFurnaceBlockEntity(BlockPos pos, BlockState state) {
+        super(ModBlockEntities.SPACE_TIME_FURNACE_BLOCK_ENTITY, pos, state, RecipeType.SMELTING);
+    }
+
+    @Override
+    protected Text getContainerName() {
+        return Text.translatable("block.divinityunbound.space_time_furnace");
+    }
+
+    @Override
+    protected ScreenHandler createScreenHandler(int syncId, PlayerInventory playerInventory) {
+        return new SpaceTimeFurnaceScreenHandler(syncId, playerInventory, this, this.propertyDelegate);
+    }
+
+    @Override
+    protected int getFuelTime(ItemStack fuel) {
+        return super.getFuelTime(fuel) / 2;
+    }
+
+
+}